feat: complete foundation milestone 8

+++ src/Microsoft/Xna/Framework/Input/__init__.pyi
@@ -432,7 +432,7 @@ class Mouse:
     def GetState() -> MouseState: ...
     @staticmethod
     def SetPosition(x: int, y: int) -> None: ...
-    WindowHandle: ClassVar[IntPtr]
+    WindowHandle: ClassVar[int]
 
 class MouseState:
     @overload
